float64 -> int64

+++ solrman/solrmanapi/cloudstatus.go
@@ -43,7 +43,7 @@ type SolrCoreStatus struct {
 	HasStats        bool      // if false, core status could not be queried and following attributes are expected to be -1
 	NumDocs         int64     // total number of indexed documents
 	IndexSize       int64     // in bytes
-	OnDiskIndexSize float64   // in bytes, differs from IndexSize for TeeDirectory implementations
+	OnDiskIndexSize int64     // in bytes, differs from IndexSize for TeeDirectory implementations
 	Type            string    // e.g. "NRT", "PULL", "TLOG"
 	Version         int64     // Update version of the replica (may be empty)
 	LastModified    time.Time // Last time the replica was modified (may be empty)
